피드백 반영 후 수정

+++ app/src/main/java/com/example/calculator_re/AbstractOperation.kt
@@ -1,5 +1,5 @@
 package com.example.calculator_re
 
 abstract class AbstractOperation {
-    abstract fun operate(num1: Int, num2: Int): Double
+    abstract fun operate(num1: Double, num2: Double): Double
 }
