+++ app/src/main/java/com/example/calculator_re/AddOperation.kt
@@ -2,6 +2,6 @@ package com.example.calculator_re
 
 
 class AddOperation: AbstractOperation() {
-    override fun operate(num1: Int, num2: Int): Double = (num1 + num2).toDouble()
+    override fun operate(num1: Double, num2: Double): Double = (num1 + num2)
 
 }
