+++ app/src/main/java/com/example/calculator_re/SubstractOperation.kt
@@ -2,5 +2,5 @@ package com.example.calculator_re
 
 
 class SubstractOperation: AbstractOperation() {
-    override fun operate(num1: Int, num2: Int): Double = (num1 - num2).toDouble()
+    override fun operate(num1: Double, num2: Double): Double = (num1 - num2)
 }
